binforge/requirements.txt
pytest
PyYAML>=6.0
colorama>=0.4.6